Fix bug that really slowed down the engine.
Problem is that if the creature's speed is low and the player is out of visible contact with any monster generators then the engine can spend a long time generating and discarding and doing
visibility tests on hypothetical monsters that will never interact with the player.  The engine now generates about ten monsters per round at random, which is few enough that it won't normally be a
problem.